% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_plyr.R
\name{llply_par}
\alias{llply_par}
\alias{parLapply2}
\alias{apply_par}
\title{parallel apply and llply}
\usage{
llply_par(X, FUN, ..., byrow = TRUE, .combine = c)

parLapply2(X, FUN, ..., byrow = TRUE, .combine = c)

apply_par(X, .margins = 1, FUN, ..., .progress = "text")
}
\description{
parallel apply and llply
}
\keyword{internal}
