% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_rast.R
\name{slope_arr}
\alias{slope_arr}
\title{slope_arr}
\usage{
slope_arr(arr, fun = rtrend::slope_mk, return.list = FALSE)
}
\value{
t, A 3d array, with the dim of \verb{[nx, ny, 2]}.
\itemize{
\item \code{t[,,1]}: slope
\item \code{t[,,2]}: pvalue
}
}
\description{
slope_arr
}
