% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_rast.R
\name{slope_rast}
\alias{slope_rast}
\alias{rast_filter_time}
\title{calculate slope of rast object}
\usage{
slope_rast(
  r,
  period = c(2001, 2020),
  outfile = NULL,
  overwrite = FALSE,
  ...,
  fun = rtrend::slope_mk
)

rast_filter_time(r, period = c(2001, 2020))
}
\arguments{
\item{r}{A yearly rast object, which should have time attribute}

\item{period}{\code{c(year_begin, year_end)}}

\item{outfile}{The path of outputed tiff file. If specified, \code{slope} and
\code{pvalue} will be written into \code{outfile}.}

\item{overwrite}{logical. If \code{TRUE}, \code{outfile} is overwritten.}

\item{...}{other parameters ignored}

\item{fun}{the function used to calculate slope, see \code{\link[=slope]{slope()}} for details.}
}
\value{
A terra rast object, with bands of \code{slope} and \code{pvalue}.
}
\description{
calculate slope of rast object
}
\seealso{
\code{\link[terra:rast]{terra::rast()}}
}
