% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_plyr.R
\name{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(x, nchunk = 6, byrow = TRUE)
}
\arguments{
\item{x}{a vector, list, or matrix (not support 3d array)}

\item{nchunk}{the number of chunks to be splitted}

\item{byrow}{If \code{TRUE}, split by row, otherwise by column.}
}
\description{
split_data
}
\keyword{internal}
