% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_rast.R
\name{slope_rast}
\alias{slope_rast}
\title{calculate MK slope of rast object}
\usage{
slope_rast(
  r,
  period = c(2001, 2020),
  fun = rtrend::slope_mk,
  outfile = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{r}{A rast object}

\item{...}{ignored}
}
\description{
calculate MK slope of rast object
}
\seealso{
\code{\link[terra:rast]{terra::rast()}}
}
